+++ packages/cli-client/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.0.65](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.64...@wireapp/cli-client@12.0.65) (2023-02-08)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.64](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.63...@wireapp/cli-client@12.0.64) (2023-02-07)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.63](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.62...@wireapp/cli-client@12.0.63) (2023-02-07)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.62](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.61...@wireapp/cli-client@12.0.62) (2023-02-07)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.61](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.60...@wireapp/cli-client@12.0.61) (2023-02-06)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.60](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.59...@wireapp/cli-client@12.0.60) (2023-02-06)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.59](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.58...@wireapp/cli-client@12.0.59) (2023-02-02)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.58](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.57...@wireapp/cli-client@12.0.58) (2023-02-02)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.57](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.56...@wireapp/cli-client@12.0.57) (2023-02-01)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
 ## [12.0.56](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.55...@wireapp/cli-client@12.0.56) (2023-01-31)
 
 **Note:** Version bump only for package @wireapp/cli-client